requirements.txt
python-telegram-bot==20.8
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
httpx==0.27.0
emoji==2.10.0
nltk==3.8.1
